cycles_backend/requirements.txt
aiofiles==23.1.0
aioredis==2.0.1
apple-music-python==1.0.5
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.6
django-allauth==0.52.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-ipware==4.0.2
django-model-utils==4.3.1
django-request==1.6.2
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-multiple-models==2.1.3
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
docker==6.0.1
fb-idb==1.1.7
firebase-admin==6.1.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
grpclib==0.4.3
h2==4.1.0
hiredis==2.2.2
hpack==4.0.0
httplib2==0.21.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonfield==3.1.0
msgpack==1.0.4
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyngrok==5.2.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
spotipy==2.22.1
sqlparse==0.4.3
swapper==1.3.0
toml==0.10.2
tomli==2.0.1
treelib==1.6.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
websocket-client==1.5.1
zope.interface==5.5.2
